refactor(machina): move cmdMachina to its own module
cmdMachina lived inline in the 3000-line index.ts. Move it (and its local
parse helper) verbatim into src/machina.ts and import it, mirroring the
existing command-module pattern (cmdPlugin/./plugin.js, cmdClip/./clipper.js,
cmdOperate/./operate.js, cmdOpenshell/./openshell-cli.js).

Behavior-preserving — identical body; the dispatch switch is unchanged. Drops
the now-unused validateConnectBundle import from index.ts. -113 lines from
index.ts.
